requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.3.3
gunicorn==23.0.0
python-dotenv==1.0.1